Email refactor.
