+++ src/test/java/org/broadinstitute/consent/http/service/EmailServiceTest.java
@@ -1,21 +1,16 @@
 package org.broadinstitute.consent.http.service;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import ch.qos.logback.core.read.ListAppender;
 import com.sendgrid.Response;
 import com.sendgrid.helpers.mail.Mail;
 import freemarker.template.Template;
@@ -25,30 +20,18 @@ import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.RandomUtils;
 import org.broadinstitute.consent.http.configurations.ConsentConfiguration;
 import org.broadinstitute.consent.http.configurations.MailConfiguration;
 import org.broadinstitute.consent.http.configurations.ServicesConfiguration;
-import org.broadinstitute.consent.http.db.DacDAO;
-import org.broadinstitute.consent.http.db.DarCollectionDAO;
-import org.broadinstitute.consent.http.db.DataAccessRequestDAO;
-import org.broadinstitute.consent.http.db.DatasetDAO;
-import org.broadinstitute.consent.http.db.ElectionDAO;
 import org.broadinstitute.consent.http.db.MailMessageDAO;
 import org.broadinstitute.consent.http.db.UserDAO;
-import org.broadinstitute.consent.http.db.VoteDAO;
-import org.broadinstitute.consent.http.enumeration.ElectionType;
 import org.broadinstitute.consent.http.enumeration.EmailType;
 import org.broadinstitute.consent.http.mail.SendGridAPI;
 import org.broadinstitute.consent.http.mail.freemarker.FreeMarkerTemplateHelper;
 import org.broadinstitute.consent.http.models.Dac;
-import org.broadinstitute.consent.http.models.DarCollection;
-import org.broadinstitute.consent.http.models.DataAccessRequest;
-import org.broadinstitute.consent.http.models.Election;
 import org.broadinstitute.consent.http.models.User;
-import org.broadinstitute.consent.http.models.Vote;
 import org.broadinstitute.consent.http.models.mail.MailMessage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -56,7 +39,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.slf4j.LoggerFactory;
 
 /**
  * This class can be used to functionally test email notifications as well as unit test. To enable
@@ -69,21 +51,9 @@ class EmailServiceTest {
 
   private EmailService service;
   @Mock
-  private DarCollectionDAO collectionDAO;
-  @Mock
-  private VoteDAO voteDAO;
-  @Mock
-  private ElectionDAO electionDAO;
-  @Mock
   private UserDAO userDAO;
   @Mock
   private MailMessageDAO emailDAO;
-  @Mock
-  private DatasetDAO datasetDAO;
-  @Mock
-  private DacDAO dacDAO;
-  @Mock
-  private DataAccessRequestDAO dataAccessRequestDAO;
   @Mock
   private SendGridAPI sendGridAPI;
 
@@ -102,14 +72,8 @@ class EmailServiceTest {
     ServicesConfiguration servicesConfiguration = config.getServicesConfiguration();
     servicesConfiguration.setLocalURL(SERVER_URL);
     service = new EmailService(
-        collectionDAO,
-        voteDAO,
-        electionDAO,
         userDAO,
         emailDAO,
-        datasetDAO,
-        dacDAO,
-        dataAccessRequestDAO,
         sendGridAPI,
         templateHelper,
         config);
@@ -391,244 +355,4 @@ class EmailServiceTest {
     );
   }
 
-  @Test
-  void testSendReminderMessage() throws Exception {
-    Election election = new Election();
-    election.setElectionId(RandomUtils.nextInt());
-    election.setReferenceId(UUID.randomUUID().toString());
-    election.setElectionType(ElectionType.DATA_ACCESS.getValue());
-    when(electionDAO.findElectionWithFinalVoteById(any())).thenReturn(election);
-
-    Vote vote = new Vote();
-    vote.setVoteId(RandomUtils.nextInt());
-    vote.setElectionId(election.getElectionId());
-    when(voteDAO.findVoteById(any())).thenReturn(vote);
-
-    DarCollection collection = new DarCollection();
-    collection.setDarCollectionId(RandomUtils.nextInt());
-    collection.setDarCode("DAR-12345");
-    when(collectionDAO.findDARCollectionByReferenceId(any())).thenReturn(collection);
-
-    User user = new User();
-    user.setDisplayName(RandomStringUtils.randomAlphanumeric(10));
-    user.setEmail(RandomStringUtils.randomAlphanumeric(10));
-    when(userDAO.findUserById(any())).thenReturn(user);
-
-    when(templateHelper.getTemplate(EmailType.REMINDER.templateName)).thenReturn(mock());
-
-    service.sendReminderMessage(vote.getVoteId());
-    verify(sendGridAPI).sendMessage(any(), any());
-    verify(emailDAO)
-        .insert(
-            eq(String.valueOf(vote.getElectionId())),
-            eq(vote.getVoteId()),
-            eq(user.getUserId()),
-            eq(EmailType.REMINDER.getTypeInt()),
-            any(),
-            any(),
-            any(),
-            any(),
-            any());
-    verify(voteDAO).updateVoteReminderFlag(vote.getVoteId(), true);
-  }
-
-  @Test
-  void sendDarExpiredMessage() throws Exception {
-    User user = new User();
-    user.setUserId(123);
-    user.setDisplayName("John Doe");
-    user.setEmail("jd@somewhere");
-    String darCode = "DAR-12345";
-    Integer otherUserId = 456;
-    String referenceId = UUID.randomUUID().toString();
-    when(templateHelper.getTemplate(EmailType.DAR_EXPIRED.templateName)).thenReturn(mock());
-
-    service.sendDarExpiredMessage(user, darCode, otherUserId, referenceId);
-    verify(sendGridAPI).sendMessage(any(), eq(user.getEmail()));
-    verify(emailDAO).insert(
-        eq(referenceId),
-        isNull(),
-        eq(otherUserId),
-        eq(EmailType.DAR_EXPIRED.getTypeInt()),
-        any(),
-        any(),
-        any(),
-        any(),
-        any());
-  }
-
-  @Test
-  void sendDarExpirationReminderMessage() throws Exception {
-    User user = new User();
-    user.setUserId(123);
-    user.setDisplayName("John Doe");
-    user.setEmail("jd@somewhere");
-    String darCode = "DAR-12345";
-    String referenceId = UUID.randomUUID().toString();
-    Integer otherUserId = 456;
-    when(templateHelper.getTemplate(EmailType.DAR_EXPIRATION_REMINDER.templateName)).thenReturn(
-        mock());
-
-    service.sendDarExpirationReminderMessage(user, darCode, otherUserId, referenceId);
-    verify(sendGridAPI).sendMessage(any(), eq(user.getEmail()));
-    verify(emailDAO).insert(
-        eq(referenceId),
-        isNull(),
-        eq(otherUserId),
-        eq(EmailType.DAR_EXPIRATION_REMINDER.getTypeInt()),
-        any(),
-        any(),
-        any(),
-        any(),
-        any());
-  }
-
-  @Test
-  void sendExpirationNoticesTest() throws IOException {
-    User user1 = new User();
-    user1.setUserId(123);
-    user1.setDisplayName("John Doe");
-    user1.setEmail("jd@somewhere");
-
-    User user2 = new User();
-    user2.setUserId(124);
-    user2.setDisplayName("Jane Doe");
-    user2.setEmail("jd@somewhereelse");
-
-    DataAccessRequest dar1 = getMockedDar("DAR-12345", UUID.randomUUID().toString(), user1);
-    DataAccessRequest dar2 = getMockedDar("DAR-12346", UUID.randomUUID().toString(), user2);
-
-    when(userDAO.findUserById(user1.getUserId())).thenReturn(user1);
-    when(userDAO.findUserById(user2.getUserId())).thenReturn(user2);
-    List<DataAccessRequest> dars = List.of(dar1, dar2);
-    when(dataAccessRequestDAO.findAgedDARsByEmailTypeOlderThanInterval(any(), any(), any())).thenReturn(dars);
-    when(templateHelper.getTemplate(EmailType.DAR_EXPIRATION_REMINDER.templateName)).thenReturn(mock());
-    when(templateHelper.getTemplate(EmailType.DAR_EXPIRED.templateName)).thenReturn(mock());
-
-    assertDoesNotThrow(() -> service.sendExpirationNotices());
-    verify(emailDAO)
-        .insert(
-            eq(dar1.getReferenceId()),
-            isNull(),
-            eq(user1.getUserId()),
-            eq(EmailType.DAR_EXPIRED.getTypeInt()),
-            any(),
-            any(),
-            any(),
-            any(),
-            any());
-    verify(emailDAO)
-        .insert(
-            eq(dar1.getReferenceId()),
-            isNull(),
-            eq(user1.getUserId()),
-            eq(EmailType.DAR_EXPIRATION_REMINDER.getTypeInt()),
-            any(),
-            any(),
-            any(),
-            any(),
-            any());
-    verify(emailDAO)
-        .insert(
-            eq(dar2.getReferenceId()),
-            isNull(),
-            eq(user2.getUserId()),
-            eq(EmailType.DAR_EXPIRED.getTypeInt()),
-            any(),
-            any(),
-            any(),
-            any(),
-            any());
-    verify(emailDAO)
-        .insert(
-            eq(dar2.getReferenceId()),
-            isNull(),
-            eq(user2.getUserId()),
-            eq(EmailType.DAR_EXPIRATION_REMINDER.getTypeInt()),
-            any(),
-            any(),
-            any(),
-            any(),
-            any());
-    verify(emailDAO, times(4))
-        .insert(any(), any(), any(), any(), any(), any(), any(), any(), any());
-  }
-
-  @Test
-  void sendExpirationNoticesTestMissingEmailForOneUser() throws IOException {
-    ListAppender<ILoggingEvent> listAppender = new ListAppender<>();
-    ch.qos.logback.classic.Logger log = (ch.qos.logback.classic.Logger) LoggerFactory.getLogger(EmailService.class);
-    listAppender.start();
-    log.addAppender(listAppender);
-    User user1 = new User();
-    user1.setUserId(123);
-    user1.setDisplayName("John Doe");
-    user1.setEmail("jd@somewhere");
-
-    User user2 = new User();
-    user2.setUserId(124);
-    user2.setDisplayName("Jane Doe");
-
-    DataAccessRequest dar1 =getMockedDar("DAR-12345", UUID.randomUUID().toString(), user1);
-    DataAccessRequest dar2 = getMockedDar("DAR-12346", UUID.randomUUID().toString(), user2);
-
-    when(userDAO.findUserById(user1.getUserId())).thenReturn(user1);
-    when(userDAO.findUserById(user2.getUserId())).thenReturn(user2);
-
-    List<DataAccessRequest> dars = List.of(dar2, dar1);
-
-    when(dataAccessRequestDAO.findAgedDARsByEmailTypeOlderThanInterval(any(), any(), any())).thenReturn(dars);
-    when(templateHelper.getTemplate(EmailType.DAR_EXPIRATION_REMINDER.templateName)).thenReturn(mock());
-    when(templateHelper.getTemplate(EmailType.DAR_EXPIRED.templateName)).thenReturn(mock());
-
-    assertDoesNotThrow(()->service.sendExpirationNotices());
-
-    verify(emailDAO).insert(eq(dar1.getReferenceId()),isNull(), eq(user1.getUserId()), eq(EmailType.DAR_EXPIRED.getTypeInt()), any(), any(), any(), any(), any());
-    verify(emailDAO).insert(eq(dar1.getReferenceId()),isNull(), eq(user1.getUserId()), eq(EmailType.DAR_EXPIRATION_REMINDER.getTypeInt()), any(), any(), any(), any(), any());
-    verify(emailDAO, times(2)).insert(any(), any(), any(), any(), any(), any(), any(), any(), any());
-    assertEquals(2, listAppender.list.size());
-  }
-
-  @Test
-  void sendExpirationNoticesTestUnderlyingExceptionThrownSendingOneTypeOfMessage() throws IOException {
-    ListAppender<ILoggingEvent> listAppender = new ListAppender<>();
-    ch.qos.logback.classic.Logger log = (ch.qos.logback.classic.Logger) LoggerFactory.getLogger(EmailService.class);
-    listAppender.start();
-    log.addAppender(listAppender);
-    User user1 = new User();
-    user1.setUserId(123);
-    user1.setDisplayName("John Doe");
-    user1.setEmail("jd@somewhere");
-
-    User user2 = new User();
-    user2.setUserId(124);
-    user2.setDisplayName("Jane Doe");
-    user2.setEmail("jane@somewhere");
-
-    DataAccessRequest dar1 =getMockedDar("DAR-12345", UUID.randomUUID().toString(), user1);
-    DataAccessRequest dar2 = getMockedDar("DAR-12346", UUID.randomUUID().toString(), user2);
-
-    when(userDAO.findUserById(user1.getUserId())).thenReturn(user1);
-    when(userDAO.findUserById(user2.getUserId())).thenReturn(user2);
-
-    List<DataAccessRequest> dars = List.of(dar2, dar1);
-
-    when(dataAccessRequestDAO.findAgedDARsByEmailTypeOlderThanInterval(any(), any(), any())).thenReturn(dars);
-    when(templateHelper.getTemplate(EmailType.DAR_EXPIRED.templateName)).thenReturn(mock());
-
-    assertDoesNotThrow(()->service.sendExpirationNotices());
-
-    verify(emailDAO).insert(eq(dar1.getReferenceId()),isNull(), eq(user1.getUserId()), eq(EmailType.DAR_EXPIRED.getTypeInt()), any(), any(), any(), any(), any());
-    verify(emailDAO).insert(eq(dar2.getReferenceId()),isNull(), eq(user2.getUserId()), eq(EmailType.DAR_EXPIRED.getTypeInt()), any(), any(), any(), any(), any());
-    verify(emailDAO, times(2)).insert(any(), any(), any(), any(), any(), any(), any(), any(), any());
-    assertEquals(2, listAppender.list.size());
-  }
-
-  private DataAccessRequest getMockedDar(String darCode, String referenceId, User user) {
-    DataAccessRequest dar = mock(DataAccessRequest.class);
-    when(dar.getReferenceId()).thenReturn(referenceId);
-    when(dar.getDarCode()).thenReturn(darCode);
-    when(dar.getUserId()).thenReturn(user.getUserId());
-    return dar;
-  }
 }
